+++ src/ClientParent.php
@@ -3,11 +3,12 @@
 namespace Bayfront\MultiCurl;
 
 use Bayfront\ArrayHelpers\Arr;
+use Bayfront\MultiCurl\Exceptions\ClientException;
 
 class ClientParent
 {
 
-    private const MULTI_CURL_VERSION = '2.2.0';
+    private const MULTI_CURL_VERSION = '3.0.0';
 
     private string $base_url;