+++ CHANGELOG.md
@@ -12,6 +12,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities
 
+## [3.0.0] - 2025.01.15
+
+### Added
+
+- Added `InteractsWithApi` trait.
+
+### Changed
+
+- Minimum PHP requirement is now `^8.2`.
+- Moved all exceptions to the `Bayfront\MultiCurl\Exceptions` namespace.
+- Updated documentation.
+ 
 ## [2.2.1] - 2024.12.23
 
 ### Added